Seperate state machines
